Updated additional build scripts to not attempt download if "buildsrc" directory is being used
FossilOrigin-Name: d860d314ddd399e0e91c888e27519d79663f0ae0

+++ mk4tcl/build.sh
@@ -32,9 +32,11 @@ export TCL_VERSION
 if [ ! -f "${SRC}" ]; then
 	mkdir 'src' >/dev/null 2>/dev/null
 
-	rm -f "${SRC}.tmp"
-	wget -O "${SRC}.tmp" "${SRCURL}" || exit 1
-	mv "${SRC}.tmp" "${SRC}"
+	if [ ! -d 'buildsrc' ]; then
+		rm -f "${SRC}.tmp"
+		wget -O "${SRC}.tmp" "${SRCURL}" || exit 1
+		mv "${SRC}.tmp" "${SRC}"
+	fi
 fi
 
 (
